requirements.txt
Flask==1.1.1
line-bot-sdk==1.18.0
beautifulsoup4==4.8.2
urllib3==1.25.8
soupsieve==1.9.5
pandas==1.1.0
#mysqlclient==1.3.12
#oauth2client==4.1.3
#gspread==3.1.0
#pydrive==1.3.1